birthdayをint型に変更1

+++ app/src/main/java/org/t_robop/y_ogawara/tancolle/Data.java
@@ -74,7 +74,7 @@ public class Data {
         this.kana = kana;
     }
 
-    public long getBirthday() {
+    public int getBirthday() {
         return birthday;
     }
